% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassicRNG.R
\name{allRNG}
\alias{allRNG}
\title{Get all classical RNG measures}
\usage{
allRNG(y, minScale = NA, maxScale = NA, R = TRUE, RNG = TRUE,
  RNG2 = TRUE, RF = TRUE, Coupon = TRUE, NSQ = TRUE,
  Adjacency = TRUE, TPI = TRUE, PhaseLength = TRUE, Runs = TRUE,
  RepDistance = TRUE, RepGap = TRUE, Phi = TRUE)
}
\arguments{
\item{y}{A sequence of symbols. If \code{y} is non-numeric, unique elements will be labelled by an integer value.}

\item{minScale}{Minimum expected value. If \code{y} is a character vector this should refer to the lowest numeric code used.}

\item{maxScale}{Maximum expected value. If \code{y} is a character vector this should refer to the highest numeric code used.}

\item{R}{Redundancy (default = \code{TRUE})}

\item{RNG}{RNG (default = \code{TRUE})}

\item{RNG2}{RNG2 (default = \code{TRUE})}

\item{RF}{Response frequencies (default = \code{TRUE})}

\item{Coupon}{Coupon (default = \code{TRUE})}

\item{NSQ}{NSQ (default = \code{TRUE})}

\item{Adjacency}{Adjacency (default = \code{TRUE})}

\item{TPI}{TPI (default = \code{TRUE})}

\item{PhaseLength}{PhaseLength (default = \code{TRUE})}

\item{Runs}{Runs (default = \code{TRUE})}

\item{RepDistance}{Repetition Distance (default = \code{TRUE})}

\item{RepGap}{Repetition Gap (default = \code{TRUE})}

\item{Phi}{Phi index (default = \code{TRUE})}
}
\value{
A list object of classical RNG task measures.
}
\description{
Get all classical RNG measures
}
\examples{

y <- round(runif(100,1,9))

# Omit the Frequecny tables
allRNG(y, minScale = 1, maxScale = 9, RF = FALSE)

}
