% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassicRNG.R
\name{check_y}
\alias{check_y}
\title{Check vector}
\usage{
check_y(y, minScale = NA, maxScale = NA, noZero = FALSE,
  toNumeric = TRUE, responseAlternatives = NA)
}
\arguments{
\item{y}{A vector. If not a numeric vector and \code{toNumeric = TRUE}, unique elements will be labelled by an integer value.}

\item{minScale}{Minimum expected value. If \code{y} is a character vector this should refer to the lowest numeric code used.}

\item{maxScale}{Maximum expected value. If \code{y} is a character vector this should refer to the highest numeric code used.}

\item{noZero}{Can the series have value \code{0}?}

\item{toNumeric}{If \code{TRUE}, the unique elements in a non-numeric vector will be labelled by an integer value.}

\item{responseAlternatives}{An optional vector of possible response alternatives. If \code{NA}, \code{responseAlternatives} will be set to \code{seq(minScale,maxScale)}}
}
\value{
Trasformed \code{y} (if necessary), or, an error message.
}
\description{
Check vector
}
\examples{

y <- letters
check_y(y)

}
