% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassicRNG.R
\name{classicalRNG}
\alias{classicalRNG}
\alias{Redundancy}
\alias{RNG}
\alias{RNG2}
\alias{RF}
\alias{Coupon}
\alias{NSQ}
\alias{FOD}
\alias{Adjacency}
\alias{TPI}
\alias{PhL}
\alias{Runs}
\alias{repDistance}
\alias{repGap}
\alias{phiIndex}
\title{Measures of the classical RNG task}
\usage{
Redundancy(y, minScale = NA, maxScale = NA)

RNG(y, minScale = NA, maxScale = NA, results = c("classical", "randseqR")[2])

RNG2(y, minScale = NA, maxScale = NA, results = c("classical", "randseqR")[2])

RF(y, minScale = NA, maxScale = NA)

Coupon(y, minScale = NA, maxScale = NA)

NSQ(y, minScale = NA, maxScale = NA, results = c("classical", "randseqR")[2])

FOD(y, minScale = NA, maxScale = NA, results = c("classical", "randseqR")[2])

Adjacency(
  y,
  minScale = NA,
  maxScale = NA,
  results = c("classical", "randseqR")[2]
)

TPI(y, minScale = NA, maxScale = NA)

PhL(y, minScale = NA, maxScale = NA)

Runs(y, minScale = NA, maxScale = NA, results = c("classical", "randseqR")[2])

repDistance(y, minScale = NA, maxScale = NA)

repGap(y, minScale = NA, maxScale = NA)

phiIndex(
  y,
  minScale = NA,
  maxScale = NA,
  responseAlternatives = NA,
  maxOrder = 7
)
}
\arguments{
\item{y}{A sequence of symbols. If \code{y} is non-numeric, unique elements will be labelled by an integer value.}

\item{minScale}{Minimum expected value. If \code{y} is a character vector this should refer to the lowest numeric code used.}

\item{maxScale}{Maximum expected value. If \code{y} is a character vector this should refer to the highest numeric code used.}

\item{results}{either 'randseqR' (default) or 'classical. randseqR gives a better consistency among RNG measures, while classical gives output comparible to RGcalc by Towse and Neil (1998)}

\item{responseAlternatives}{An optional vector of possible response alternatives. If \code{NA}, \code{responseAlternatives} will be set to \code{seq(minScale,maxScale)}}

\item{maxOrder}{}
}
\value{
Output
}
\description{
Individual functions to get the measures of the classical RNG task. Use function \code{allRNG()} to get a list with all (or selected) measures.
}
\examples{

y <- round(runif(100,1,9))

R(y, maxScale = 9)

}
\seealso{
\code{\link[=allRNG]{allRNG()}} to get \emph{all}, or, a selected list of measures.
}
