% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassicRNG.R
\name{findTPIs}
\alias{findTPIs}
\title{Find turning points}
\usage{
findTPIs(y)
}
\arguments{
\item{y}{Numeric vector.}
}
\value{
A list object with \code{start} and \code{end} indices of turning points.
}
\description{
Find turning points
}
