% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassicRNG.R
\name{first_occurence}
\alias{first_occurence}
\title{Find first occurrence}
\usage{
first_occurence(y, minScale, maxScale)
}
\arguments{
\item{y}{Numeric vector}

\item{minScale}{Minimum expected value.}

\item{maxScale}{Maximum expected value.}
}
\value{
vector with ocurrences
}
\description{
Find first occurrence in vector \code{minScale:maxScale}
}
\keyword{internal}
