% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassicRNG.R
\name{observed_expected}
\alias{observed_expected}
\title{Observed and expected combination pairs}
\usage{
observed_expected(y, minScale, maxScale, lag)
}
\arguments{
\item{y}{Numeric vector}

\item{minScale}{Minimum expected value.}

\item{maxScale}{Maximum expected value.}

\item{lag}{Lag the series}
}
\value{
list with observed and expected paired combinations
}
\description{
Observed and expected combination pairs
}
\keyword{internal}
