% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNGseries-package.R
\docType{package}
\name{RNGseries-package}
\alias{RNGseries}
\alias{RNGseries-package}
\title{RNGseries: Functions to analyse random number sequences}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Fred Hasselman \email{f.hasselman@pwo.ru.nl} (0000-0003-1384-8361)

Other contributors:
\itemize{
  \item Wouter Oomens \email{w.oomens@donders.ru.nl} [contributor]
}

}
\keyword{internal}
