% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassicRNG.R
\name{TPIhelper}
\alias{TPIhelper}
\title{Find peaks/troughs}
\usage{
TPIhelper(y)
}
\arguments{
\item{y}{Numeric vector}
}
\value{
converted y with values 1 (trough); 2 (decrease); 3 (same); 5 (increase); 5 (peak)
}
\description{
Find peaks/troughs
}
