% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassicRNG.R
\name{classicalRNG}
\alias{classicalRNG}
\alias{Redundancy}
\alias{RNG}
\alias{RNG2}
\alias{RF}
\alias{Coupon}
\alias{NSQ}
\alias{Adjacency}
\alias{TPI}
\alias{PhL}
\alias{Runs}
\alias{repDistance}
\alias{repGap}
\alias{PhiIndex}
\title{Measures of the classical RNG task}
\usage{
Redundancy(y, maxScale = NA)

RNG(y, minScale = NA, maxScale = NA)

RNG2(y, minScale = NA, maxScale = NA)

RF(y)

Coupon(y, minScale = NA, maxScale = NA)

NSQ(y, minScale = NA, maxScale = NA)

Adjacency(y)

TPI(y)

PhL(y)

Runs(y)

repDistance(y, minScale = NA, maxscale = NA)

repGap(y, minScale = NA, maxScale = NA)

PhiIndex(y)
}
\arguments{
\item{y}{A sequence of symbols. If \code{y} is non-numeric, unique elements will be labelled by an integer value.}

\item{maxScale}{Maximum expected value. If \code{y} is a character vector this should refer to the highest numeric code used.}

\item{minScale}{Minimum expected value. If \code{y} is a character vector this should refer to the lowest numeric code used.}
}
\value{
Output
}
\description{
Individual funcations to get the measures of the classical RNG task. Use function \code{allRNG()} to get a list with all (or selected) measures.
}
\examples{

y <- round(runif(100,1,9))

R(y, maxScale = 9)

}
\seealso{
\code{\link[=allRNG]{allRNG()}} to get \emph{all}, or, a selected list of measures.
}
